Project modeling (#1)
- Split into separate packages: client and services.
- Each service will hold it's own directory and package.
- Split client into logical components.
